+++ src/entities/Alert.ts
